refactor: lint limpo no CI, teste NCM honesto e move codes.js para services
- CI: step de ESLint no workflow E2E (lint nunca rodava — 43 erros passaram despercebidos)
- lint: mapUfToUfCode (42x prettier), weather.js, cpf (console.log redundante), cep-error-handler (disable scoped)
- teste NCM: expect.assertions + comportamento real (200 [] pós-984c956) — antes passava sem validar nada
- move pages/api/indices/codes.js → services/indices/codes.js (módulo de constantes virava rota 500 em produção)
