+++ pages/api/indices/igpm/v1/intervalo/index.js
@@ -4,7 +4,7 @@ import BaseError from '@/errors/BaseError';
 import InternalError from '@/errors/InternalError';
 import { isAfter, isBefore, isValidDate, parseToDate } from '@/services/date';
 import { getIgpmByPeriod } from '@/services/indices/igpm';
-import { bcbSgsCodes } from '../../../codes';
+import { bcbSgsCodes } from '@/services/indices/codes';
 
 const action = async (request, response) => {
   try {
