+++ pages/api/indices/igpm/v1/ultimos/index.js
@@ -3,7 +3,7 @@ import BadRequestError from '@/errors/BadRequestError';
 import InternalError from '@/errors/InternalError';
 import BaseError from '@/errors/BaseError';
 import { getIgpmByLastNRecords } from '@/services/indices/igpm';
-import { bcbSgsCodes } from '../../../codes';
+import { bcbSgsCodes } from '@/services/indices/codes';
 
 const action = async (request, response) => {
   try {
