+++ util/mapUfToUfCode.js
@@ -1,34 +1,34 @@
-import NotFoundError from "@/errors/NotFoundError";
+import NotFoundError from '@/errors/NotFoundError';
 
 export const UF_CODE_MAPPER = {
-  "RO": 11,
-  "AC": 12,
-  "AM": 13,
-  "RR": 14,
-  "PA": 15,
-  "AP": 16,
-  "TO": 17,
-  "MA": 21,
-  "PI": 22,
-  "CE": 23,
-  "RN": 24,
-  "PB": 25,
-  "PE": 26,
-  "AL": 27,
-  "SE": 28,
-  "BA": 29,
-  "MG": 31,
-  "ES": 32,
-  "RJ": 33,
-  "SP": 35,
-  "PR": 41,
-  "SC": 42,
-  "RS": 43,
-  "MS": 50,
-  "MT": 51,
-  "GO": 52,
-  "DF": 53,
-}
+  RO: 11,
+  AC: 12,
+  AM: 13,
+  RR: 14,
+  PA: 15,
+  AP: 16,
+  TO: 17,
+  MA: 21,
+  PI: 22,
+  CE: 23,
+  RN: 24,
+  PB: 25,
+  PE: 26,
+  AL: 27,
+  SE: 28,
+  BA: 29,
+  MG: 31,
+  ES: 32,
+  RJ: 33,
+  SP: 35,
+  PR: 41,
+  SC: 42,
+  RS: 43,
+  MS: 50,
+  MT: 51,
+  GO: 52,
+  DF: 53,
+};
 
 /**
  * Map UF (e.g. SP) to UF Code Id on IBGE API
@@ -36,19 +36,19 @@ export const UF_CODE_MAPPER = {
  * @returns {number} UF Code
  */
 export default function mapUfToUfCode(uf) {
-    const normalizedUf = uf.toUpperCase();
+  const normalizedUf = uf.toUpperCase();
 
-    if (/^\d+$/.test(normalizedUf)) {
-        const ufCode = Number(normalizedUf);
-        if (Object.values(UF_CODE_MAPPER).includes(ufCode)) {
-            return ufCode;
-        }
+  if (/^\d+$/.test(normalizedUf)) {
+    const ufCode = Number(normalizedUf);
+    if (Object.values(UF_CODE_MAPPER).includes(ufCode)) {
+      return ufCode;
     }
+  }
 
-    let ufCode = UF_CODE_MAPPER[normalizedUf];
-    if(!ufCode){
-        throw new NotFoundError(`UF ${normalizedUf} não encontrado`);
-    }
+  const ufCode = UF_CODE_MAPPER[normalizedUf];
+  if (!ufCode) {
+    throw new NotFoundError(`UF ${normalizedUf} não encontrado`);
+  }
 
-    return ufCode;
+  return ufCode;
 }
